Trocando status das sugestoes na tabela

+++ Speckoz.Quiz/Quiz.API/Controllers/SuggestionsController.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Quiz.API.Models;
 using Quiz.API.Repository.Interfaces;
+using Quiz.Dependencies.Enums;
 
 namespace Quiz.API.Controllers
 {
@@ -12,18 +15,32 @@ namespace Quiz.API.Controllers
     public class SuggestionsController : ControllerBase
     {
         private readonly IQuestionSuggestionRepository _questionSuggestionRepository;
+        private readonly IQuestionsStatusRepository _questionsStatusRepository;
 
-        public SuggestionsController(IQuestionSuggestionRepository questionSuggestionRepository) =>
+        public SuggestionsController(IQuestionSuggestionRepository questionSuggestionRepository, 
+            IQuestionsStatusRepository questionsStatusRepository)
+        {
             _questionSuggestionRepository = questionSuggestionRepository;
-        
+            _questionsStatusRepository = questionsStatusRepository;
+        }
+
 
         // POST: /suggestions
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> CreateNewSuggestion(QuestionSuggestionModel question)
         {
             if (ModelState.IsValid)
             {
+                string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
                 var suggestion = await _questionSuggestionRepository.CreateTaskAync(question);
+
+                await _questionsStatusRepository.CreateTaskAync(new QuestionsStatusModel
+                {
+                    QuestionID = question.QuestionSuggestionID,
+                    QuestionStatus = QuestionStatusEnum.Pending,
+                    UserID = int.Parse(userId)
+                });
                 return Created($"/suggestions/{suggestion.QuestionSuggestionID}", suggestion);
             }
             return BadRequest();
@@ -38,11 +55,16 @@ namespace Quiz.API.Controllers
             return Ok(suggestions);
         }
 
-
         // DELETE: /suggestions/2
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteSuggestion(int id)
         {
+            // Coloca o status denied na tabela de status
+            var suggestion = await _questionsStatusRepository.FindByIdTaskAsync(id);
+            if (suggestion == null) return NoContent();
+            suggestion.QuestionStatus = QuestionStatusEnum.Denied;
+            await _questionsStatusRepository.UpdateTaskAync(suggestion);
+
             await _questionSuggestionRepository.DeleteSuggestionTaskAsync(id);
             return NoContent();
         }
@@ -53,6 +75,12 @@ namespace Quiz.API.Controllers
         {
             try
             {
+                // Coloca o status approved na tabela de status
+                var suggestion = await _questionsStatusRepository.FindByIdTaskAsync(id);
+                if (suggestion == null) return BadRequest();
+                suggestion.QuestionStatus = QuestionStatusEnum.Approved;
+                await _questionsStatusRepository.UpdateTaskAync(suggestion);
+
                 await _questionSuggestionRepository.ApproveSuggestion(id);
                 return Ok();
             }
@@ -61,5 +89,10 @@ namespace Quiz.API.Controllers
                 return BadRequest();
             }
         }
+
+        // GET: /suggestions/status
+        [HttpGet("status")]
+        public async Task<IActionResult> GetStatus() =>
+            Ok(await _questionsStatusRepository.GetStatus());
     }
 }
